Refactor admin routes into guarded layout
